requirements.txt
bcrypt==5.0.0
beautifulsoup4==4.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
html5lib==1.1
idna==3.10
lxml==6.0.2
numpy==2.2.6
pandas==2.3.2
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
scipy==1.15.3
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
